Rename `classes` -> `plasma` (#842)
