+++ changelog/842.breaking.rst
@@ -1,0 +1,1 @@
+Renamed the `plasmapy.classes` subpackage to `plasmapy.plasma`.
