+++ plasmapy/plasma/exceptions.py
