+++ plasmapy/plasma/plasma_base.py
@@ -5,7 +5,7 @@ __all__ = ["BasePlasma", "GenericPlasma"]
 
 class BasePlasma(ABC):
     """
-    Registration class for `~plasmapy.classes.GenericPlasma` and declares
+    Registration class for `~plasmapy.plasma.GenericPlasma` and declares
     some abstract methods for data common in different kinds of plasmas.
 
     This class checks for the existance of a method named ``is_datasource_for``
@@ -55,7 +55,7 @@ class BasePlasma(ABC):
 class GenericPlasma(BasePlasma):
     """
     A Generic Plasma class. This class contains definitions for abstract
-    methods declared in the `~plasmapy.classes.plasma_base.BaseClass`.
+    methods declared in the `~plasmapy.plasma.plasma_base.BaseClass`.
     """
 
     def __init__(self, **kwargs):
