+++ plasmapy/plasma/sources/plasmablob.py
@@ -4,12 +4,11 @@ Defines the core Plasma class used by PlasmaPy to represent plasma properties.
 import warnings
 
 import astropy.units as u
-
-from plasmapy.classes.plasma_base import GenericPlasma
 from plasmapy.formulary.collisions import coupling_parameter
 from plasmapy.formulary.dimensionless import quantum_theta
 from plasmapy.formulary.parameters import _grab_charge
 from plasmapy.particles import particle_mass
+from plasmapy.plasma.plasma_base import GenericPlasma
 from plasmapy.utils import CouplingWarning, call_string
 
 __all__ = ["PlasmaBlob"]
