+++ plasmapy/plasma/sources/tests/test_openpmd_hdf5.py
@@ -1,13 +1,13 @@
+from astropy import units as u
+from typing import Union, Tuple, List
 import os
 from typing import List, Tuple, Union
-
 import pytest
-from astropy import units as u
 
-import plasmapy.classes
-from plasmapy.classes.exceptions import DataStandardError
-from plasmapy.classes.sources import openpmd_hdf5
+from plasmapy.plasma.sources import openpmd_hdf5
+from plasmapy.plasma.exceptions import DataStandardError
 from plasmapy.particles.data.test import data_dir
+import plasmapy.plasma
 
 
 @pytest.fixture(scope="module")
@@ -147,5 +147,5 @@ def test_non_openpmd_hdf5():
 
 
 def test_HDF5Reader(h5_2d):
-    assert isinstance(h5_2d, plasmapy.classes.sources.HDF5Reader)
-    assert isinstance(h5_2d, plasmapy.classes.BasePlasma)
+    assert isinstance(h5_2d, plasmapy.plasma.sources.HDF5Reader)
+    assert isinstance(h5_2d, plasmapy.plasma.BasePlasma)
