+++ plasmapy/plasma/sources/tests/test_plasma3d.py
@@ -2,8 +2,8 @@ import astropy.units as u
 import numpy as np
 import pytest
 
-from plasmapy.classes.sources import plasma3d
 from plasmapy.particles.exceptions import InvalidParticleError
+from plasmapy.plasma.sources import plasma3d
 
 
 @pytest.mark.parametrize(
