+++ plasmapy/plasma/sources/tests/test_plasmablob.py
@@ -2,9 +2,9 @@ import astropy.units as u
 import numpy as np
 import pytest
 
-from plasmapy.classes.sources import plasma3d, plasmablob
 from plasmapy.formulary import magnetostatics
 from plasmapy.particles.exceptions import InvalidParticleError
+from plasmapy.plasma.sources import plasma3d, plasmablob
 from plasmapy.utils.exceptions import CouplingWarning
